test: CompanyQuoteServiceUnitTest.java 를 통합테스트로 변경
